restructure agents
