+++ experimental/agents/ack_generator_agent/tools.py
@@ -16,19 +16,17 @@ from strands import tool, Agent
 import json
 from rich.console import Console
 from ack_builder_agent.tools import build_controller, read_build_log, sleep, verify_build_completion
-from ack_builder_agent.utils.constants import ACK_SYSTEM_PROMPT as BUILDER_SYSTEM_PROMPT
-
-# Import settings and utility functions from utils package
-from ack_generator_agent.utils import (
-    settings,
+from ack_builder_agent.prompts import ACK_BUILDER_SYSTEM_PROMPT
+from utils.settings import settings
+from utils.repo import (
     ensure_ack_directories,
     ensure_service_repo_cloned,
     ensure_aws_sdk_go_v2_cloned,
 )
 
-from ack_generator_agent.utils.memory_agent import MemoryAgent
-from ack_generator_agent.utils.knowledge_base import retrieve_from_knowledge_base
-from ack_generator_agent.utils.docs_agent import DocsAgent
+from utils.memory_agent import MemoryAgent
+from utils.knowledge_base import retrieve_from_knowledge_base
+from utils.docs_agent import DocsAgent
 
 console = Console()
 memory_agent = MemoryAgent()
@@ -109,7 +107,7 @@ def build_controller_agent(service: str) -> str:
     """
     try:
         builder_agent = Agent(
-            system_prompt=BUILDER_SYSTEM_PROMPT,
+            system_prompt=ACK_BUILDER_SYSTEM_PROMPT,
             tools=[build_controller, read_build_log, sleep, verify_build_completion]
         )
         response = builder_agent(service)
@@ -148,25 +146,10 @@ def update_service_generator_config(service: str, new_generator_yaml: str) -> st
         return f"Successfully updated generator.yaml for {service} at {generator_config_path}"
     except Exception as e:
         return f"Error updating generator.yaml for {service}: {str(e)}" 
-    
-
-# Use the dedicated memory agent for error/solution management
-@tool
-def error_lookup(error_message: str) -> str:
-    """
-    Look up the error message in the agent's memory and return the corresponding solution.
-    
-    Args:
-        error_message: The error message to look up
-        
-    Returns:
-        str: Previously known solution or None if not found
-    """
-    return memory_agent.lookup_error_solution(error_message)
 
 
 @tool
-def add_memory(content: str, metadata: dict = None) -> str:
+def add_memory(content: str, metadata: dict) -> str:
     """
     Manually add a memory to the agent's memory store.
     
@@ -216,7 +199,7 @@ def lookup_code_generator_config(service: str, resource: str) -> str:
     return f""
 
 @tool
-def save_error_solution(error_message: str, solution: str) -> str:
+def save_error_solution(error_message: str, solution: str, metadata: dict) -> str:
     """
     Save an error message and its solution to the agent's memory.
     
@@ -227,7 +210,7 @@ def save_error_solution(error_message: str, solution: str) -> str:
     Returns:
         str: Success or error message
     """
-    return memory_agent.store_error_solution(error_message, solution)
+    return memory_agent.store_error_solution(error_message, solution, metadata)
 
 
 @tool
@@ -305,7 +288,7 @@ def get_aws_documentation_recommendations(url: str) -> str:
 
 
 @tool
-def find_service_documentation(service: str, resource: str = None) -> str:
+def find_service_documentation(service: str, resource: str) -> str:
     """
     Find AWS service documentation specifically for ACK controller generation.
