+++ experimental/agents/utils/memory_agent.py
@@ -1,10 +1,52 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
+# with the License. A copy of the License is located at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# or in the 'license' file accompanying this file. This file is distributed on an 'AS IS' BASIS, WITHOUT WARRANTIES
+# OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions
+# and limitations under the License.
+"""Memory agent for ACK agents."""
+
 from strands import Agent
 from strands_tools import mem0_memory
 from logging import getLogger
 from typing import Optional
-from ack_generator_agent.utils.constants import MEMORY_AGENT_SYSTEM_PROMPT
+from config.defaults import MEM0_USER_ID
 
-MEM0_USER_ID = "ack_codegen_agent_user"
+
+# Memory Agent System Prompt - specialized for ACK build error/solution management
+MEMORY_AGENT_SYSTEM_PROMPT = """You are the ACK Memory Agent, a specialized assistant that manages error-solution knowledge for AWS Controllers for Kubernetes (ACK) code generation and build processes.
+
+Your primary responsibilities:
+1. **Store Error Solutions**: When given an error and solution, store them in memory with appropriate metadata
+2. **Retrieve Solutions**: When asked about an error, search for and return relevant solutions from memory
+3. **Manage Knowledge Base**: Maintain a searchable repository of ACK-specific build errors and their fixes
+
+Key Areas of Expertise:
+- ACK code generator errors and fixes
+- generator.yaml configuration issues
+- AWS SDK Go v2 API mapping problems
+- Resource field mapping errors
+- Build process failures
+- Controller generation issues
+
+When storing memories:
+- Use descriptive metadata including error type, service, resource, and solution category
+- Tag with relevant keywords for easy retrieval
+- Include context about when and why the solution works
+
+When retrieving memories:
+- Search for semantically similar errors, not just exact matches
+- Return the most relevant and recent solutions
+- Provide context about when the solution was effective
+
+Always use user_id="ack_codegen_agent_user" for all memory operations to maintain consistency across the ACK generation workflow.
+
+Focus on being helpful, accurate, and building a comprehensive knowledge base that improves ACK controller generation over time."""
+
 
 class MemoryAgent:
     """Dedicated memory agent for ACK build error/solution management."""
@@ -17,7 +59,7 @@ class MemoryAgent:
             tools=[mem0_memory]
         )
     
-    def store_error_solution(self, error_message: str, solution: str, metadata: dict = None) -> str:
+    def store_error_solution(self, error_message: str, solution: str, metadata: dict) -> str:
         """Store an error and its solution in memory."""
         try:
             if metadata is None:
@@ -68,7 +110,7 @@ class MemoryAgent:
             self.logger.error(f"Failed to search memories: {e}")
             return f"Error searching memories: {str(e)}"
     
-    def add_knowledge(self, content: str, metadata: dict = None) -> str:
+    def add_knowledge(self, content: str, metadata: dict) -> str:
         """Add general knowledge to memory."""
         try:
             if metadata is None:
