+++ experimental/agents/utils/settings.py
@@ -8,7 +8,7 @@
 # or in the 'license' file accompanying this file. This file is distributed on an 'AS IS' BASIS, WITHOUT WARRANTIES
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions
 # and limitations under the License.
-"""Settings for ACK Generator tools."""
+"""Settings for ACK agents."""
 
 import os
 from pydantic_settings import BaseSettings
